refactor: Move session store implementation to a separate module
