+++ server/store/sessionStore.js
@@ -1,11 +1,11 @@
 import session from 'express-session';
-import pool from '../config/database.js';
+import { pool } from '../config/database.js';
 import connectPgSimple from 'connect-pg-simple';
 
 const PGStore = connectPgSimple(session);
 
 const sessionStore = new PGStore({
-  pool: pool,
+  pool: pool(),
   tableName: 'sessions',
   createTableIfMissing: true,
 });
